// src/lib2/models/Node.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Node = {
    config?: any;
    created_at?: string;
    has_pinned_data?: boolean;
    id?: string;
    label?: string;
    pipeline_id?: string;
    position?: {
        'x'?: number;
        'y'?: number;
    };
    type?: string;
};